requirements.txt
Django==5.1.7
python-docx==1.1.2
PyPDF2==3.0.1
